+++ library/src/com/viewpagerindicator/IcsLinearLayout.java
@@ -31,10 +31,10 @@ class IcsLinearLayout extends LinearLayout {
     private int mDividerPadding;
 
 
-    public IcsLinearLayout(Context context) {
+    public IcsLinearLayout(Context context, int themeAttr) {
         super(context);
 
-        TypedArray a = context.obtainStyledAttributes(null, LL, R.attr.vpiTabPageIndicatorStyle, 0);
+        TypedArray a = context.obtainStyledAttributes(null, LL, themeAttr, 0);
         setDividerDrawable(a.getDrawable(IcsLinearLayout.LL_DIVIDER));
         mDividerPadding = a.getDimensionPixelSize(LL_DIVIDER_PADDING, 0);
         mShowDividers = a.getInteger(LL_SHOW_DIVIDER, SHOW_DIVIDER_NONE);