+++ sample/src/com/viewpagerindicator/sample/TestFragmentAdapter.java
@@ -3,9 +3,16 @@ package com.viewpagerindicator.sample;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentPagerAdapter;
+import com.viewpagerindicator.IconPagerAdapter;
 
-class TestFragmentAdapter extends FragmentPagerAdapter {
+class TestFragmentAdapter extends FragmentPagerAdapter implements IconPagerAdapter {
     protected static final String[] CONTENT = new String[] { "This", "Is", "A", "Test", };
+    protected static final int[] ICONS = new int[] {
+            R.drawable.perm_group_calendar,
+            R.drawable.perm_group_camera,
+            R.drawable.perm_group_device_alarms,
+            R.drawable.perm_group_location
+    };
 
     private int mCount = CONTENT.length;
 
@@ -26,6 +33,11 @@ class TestFragmentAdapter extends FragmentPagerAdapter {
     @Override
     public CharSequence getPageTitle(int position) {
       return TestFragmentAdapter.CONTENT[position % CONTENT.length];
+    }
+
+    @Override
+    public int getIconResId(int index) {
+      return ICONS[index % ICONS.length];
     }
 
     public void setCount(int count) {