+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 Change Log
 ==========
 
-Version 2.4 *(In Development)*
+Version 2.4.0 *(In Development)*
 ------------------------------
 
+ * New `IconPageIndicator`! Uses state-list images to represent pages.
  * Support `android:background` attribute on `Canvas`-based views.
  * Title indicator allows for drawing its line, underline, and/or triangle on
    top of the titles for placement underneath a `ViewPager`.